Feature: Refactor BetSection to support lucky7B game and remove AndarBaharBetSection
