--- src/component/game/anderBahar/jsx/AndarBaharBetSection.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-import styles from "../styles/BetSection.module.css";
-
-export const AndarBaharBetSection = ({}) => {
-  return <div className={styles.andarBaharBetSection}></div>;
-};
