+++ src/component/game/common/layout/style/BetSection.module.css
@@ -4,20 +4,4 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.lowHighEvenOdd {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-}
-
-.lowHigh {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
 }
