+++ src/component/game/lucky7B/style/BetSection.module.css
@@ -1,0 +1,23 @@
+.betSection {
+  height: 100%;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.lowHighEvenOdd {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+}
+
+.lowHigh {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+}
